feat : 카테고리 삭제 #46
