+++ src/main/java/com/sparta/oneeat/category/service/StoreCategoryServiceImpl.java
@@ -6,9 +6,11 @@ import com.sparta.oneeat.category.entity.Category;
 import com.sparta.oneeat.category.repository.StoreCategotyRepository;
 import com.sparta.oneeat.common.exception.CustomException;
 import com.sparta.oneeat.common.exception.ExceptionType;
+import com.sparta.oneeat.store.repository.StoreRepository;
 import com.sparta.oneeat.user.entity.User;
 import com.sparta.oneeat.user.entity.UserRoleEnum;
 import com.sparta.oneeat.user.repository.UserRepository;
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -22,6 +24,7 @@ public class StoreCategoryServiceImpl implements StoreCategoryService{
 
     private final UserRepository userRepository;
     private final StoreCategotyRepository storeCategotyRepository;
+    private final StoreRepository storeRepository;
 
 
     // 관리자 체크 메서드
@@ -46,5 +49,23 @@ public class StoreCategoryServiceImpl implements StoreCategoryService{
         Category saved = storeCategotyRepository.save(new Category(createCategoryReqDto));
 
         return new CreateCategoryResDto(saved.getId());
+    }
+
+    @Transactional
+    public void deleteCategory(User user, UUID categoryId) {
+        checkRole(user.getRole());
+
+        //Todo 카테고리 존재 여부 검증
+        log.info("카테고리 검증 : ");
+        Category category = storeCategotyRepository.findById(categoryId).orElseThrow(() -> new CustomException(
+            ExceptionType.INTERNAL_SERVER_ERROR)); //Todo 카테고리 관련 타입으로 수정 필요함
+
+        if (!storeRepository.findByCategory(category).isEmpty()) {
+            log.info("카테고리 사용 중");
+            throw new CustomException(ExceptionType.INTERNAL_SERVER_ERROR); //Todo 카테고리 관련 타입으로 수정 필요함
+        }
+
+        //Todo 카테고리 삭제 처리
+        storeCategotyRepository.delete(category);
     }
 }
