--- src/main/java/com/sparta/oneeat/store/repository/CategoryRepository.java
@@ -1,11 +1,0 @@
-package com.sparta.oneeat.store.repository;
-
-import com.sparta.oneeat.store.entity.Category;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CategoryRepository extends JpaRepository <Category, UUID> {
-
-}
